backview/requirements.txt
numpy
pandas
fastapi
uvicorn
python-dotenv
findspark
pyspark